Renamed test package name to better match src
The test package name does not follow the src package name convention.
This is kind of silly, since then you can't test protected (package
private) methods.

Fixed to better adhere to sane java practices
