+++ src/test/java/net/starschema/clouddb/jdbc/BQResultSetFunctionTest.java
@@ -22,7 +22,7 @@
  * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
  * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
-package BQJDBC.QueryResultTest;
+package net.starschema.clouddb.jdbc;
 
 import java.sql.DriverManager;
 import java.sql.ResultSet;
@@ -30,9 +30,6 @@ import java.sql.SQLException;
 import java.sql.Statement;
 
 import junit.framework.Assert;
-import net.starschema.clouddb.jdbc.BQConnection;
-import net.starschema.clouddb.jdbc.BQSupportFuncts;
-import net.starschema.clouddb.jdbc.BQSupportMethods;
 
 import org.junit.Before;
 import org.junit.Test;
@@ -45,44 +42,43 @@ import org.slf4j.LoggerFactory;
  * @author Horváth Attila
  * @author Gunics Balazs
  */
-public class BQScrollableResultSetFunctionTest {
+public class BQResultSetFunctionTest {
 
     private static java.sql.Connection con = null;
     private static java.sql.ResultSet Result = null;
 
-    Logger logger = LoggerFactory.getLogger(BQScrollableResultSetFunctionTest.class);
+    Logger logger = LoggerFactory.getLogger(BQResultSetFunctionTest.class);
 
     @Test
     public void ChainedCursorFunctionTest() {
         this.logger.info("ChainedFunctionTest");
         try {
-            BQScrollableResultSetFunctionTest.Result.beforeFirst();
-            Assert.assertTrue(BQScrollableResultSetFunctionTest.Result.next());
-            Assert.assertEquals("you",
-                    BQScrollableResultSetFunctionTest.Result.getString(1));
+            BQResultSetFunctionTest.Result.beforeFirst();
+            Assert.assertTrue(BQResultSetFunctionTest.Result.next());
+            Assert.assertEquals("you", BQResultSetFunctionTest.Result.getString(1));
 
         } catch (SQLException e) {
             this.logger.error("SQLexception" + e.toString());
             Assert.fail("SQLException" + e.toString());
         }
         try {
-            Assert.assertTrue(BQScrollableResultSetFunctionTest.Result.absolute(10));
+            Assert.assertTrue(BQResultSetFunctionTest.Result.absolute(10));
             Assert.assertEquals("whom",
-                    BQScrollableResultSetFunctionTest.Result.getString(1));
+                    BQResultSetFunctionTest.Result.getString(1));
         } catch (SQLException e) {
             this.logger.error("SQLexception" + e.toString());
             Assert.fail("SQLException" + e.toString());
         }
 
         try {
-            Assert.assertFalse(BQScrollableResultSetFunctionTest.Result.next());
+            Assert.assertFalse(BQResultSetFunctionTest.Result.next());
         } catch (SQLException e) {
             this.logger.error("SQLexception" + e.toString());
             Assert.fail("SQLException" + e.toString());
         }
 
         try {
-            Assert.assertEquals("", BQScrollableResultSetFunctionTest.Result.getString(1));
+            Assert.assertEquals("", BQResultSetFunctionTest.Result.getString(1));
         } catch (SQLException e) {
             boolean ct = e.toString().contains(
                     "Cursor is not in a valid Position");
@@ -96,38 +92,38 @@ public class BQScrollableResultSetFunctionTest {
         }
 
         try {
-            Assert.assertTrue(BQScrollableResultSetFunctionTest.Result.first());
+            Assert.assertTrue(BQResultSetFunctionTest.Result.first());
             Assert.assertEquals("you",
-                    BQScrollableResultSetFunctionTest.Result.getString(1));
+                    BQResultSetFunctionTest.Result.getString(1));
         } catch (SQLException e) {
             this.logger.error("SQLexception" + e.toString());
             Assert.fail("SQLException" + e.toString());
         }
 
         try {
-            Assert.assertTrue(BQScrollableResultSetFunctionTest.Result.isFirst());
-            Assert.assertFalse(BQScrollableResultSetFunctionTest.Result.previous());
-            BQScrollableResultSetFunctionTest.Result.afterLast();
-            Assert.assertTrue(BQScrollableResultSetFunctionTest.Result.isAfterLast());
-            Assert.assertTrue(BQScrollableResultSetFunctionTest.Result.absolute(-1));
+            Assert.assertTrue(BQResultSetFunctionTest.Result.isFirst());
+            Assert.assertFalse(BQResultSetFunctionTest.Result.previous());
+            BQResultSetFunctionTest.Result.afterLast();
+            Assert.assertTrue(BQResultSetFunctionTest.Result.isAfterLast());
+            Assert.assertTrue(BQResultSetFunctionTest.Result.absolute(-1));
             Assert.assertEquals("whom",
-                    BQScrollableResultSetFunctionTest.Result.getString(1));
+                    BQResultSetFunctionTest.Result.getString(1));
         } catch (SQLException e) {
             this.logger.error("SQLexception" + e.toString());
             Assert.fail("SQLException" + e.toString());
         }
 
         try {
-            Assert.assertTrue(BQScrollableResultSetFunctionTest.Result.relative(-5));
+            Assert.assertTrue(BQResultSetFunctionTest.Result.relative(-5));
             Assert.assertEquals("without",
-                    BQScrollableResultSetFunctionTest.Result.getString(1));
+                    BQResultSetFunctionTest.Result.getString(1));
         } catch (SQLException e) {
             this.logger.error("SQLexception" + e.toString());
             Assert.fail("SQLException" + e.toString());
         }
 
         try {
-            Assert.assertFalse(BQScrollableResultSetFunctionTest.Result.relative(6));
+            Assert.assertFalse(BQResultSetFunctionTest.Result.relative(6));
         } catch (SQLException e) {
             this.logger.error("SQLexception" + e.toString());
             Assert.fail("SQLException" + e.toString());
@@ -135,7 +131,7 @@ public class BQScrollableResultSetFunctionTest {
 
         try {
             Assert.assertEquals("without",
-                    BQScrollableResultSetFunctionTest.Result.getString(1));
+                    BQResultSetFunctionTest.Result.getString(1));
         } catch (SQLException e) {
             boolean ct = e.toString().contains(
                     "Cursor is not in a valid Position");
@@ -213,37 +209,37 @@ public class BQScrollableResultSetFunctionTest {
     @Test
     public void isClosedValidtest() {
         try {
-            Assert.assertEquals(true, BQScrollableResultSetFunctionTest.con.isValid(0));
+            Assert.assertEquals(true, BQResultSetFunctionTest.con.isValid(0));
         } catch (SQLException e) {
             Assert.fail("Got an exception" + e.toString());
             e.printStackTrace();
         }
         try {
-            Assert.assertEquals(true, BQScrollableResultSetFunctionTest.con.isValid(10));
+            Assert.assertEquals(true, BQResultSetFunctionTest.con.isValid(10));
         } catch (SQLException e) {
             Assert.fail("Got an exception" + e.toString());
             e.printStackTrace();
         }
         try {
-            BQScrollableResultSetFunctionTest.con.isValid(-10);
+            BQResultSetFunctionTest.con.isValid(-10);
         } catch (SQLException e) {
             Assert.assertTrue(true);
             // e.printStackTrace();
         }
 
         try {
-            BQScrollableResultSetFunctionTest.con.close();
+            BQResultSetFunctionTest.con.close();
         } catch (SQLException e) {
             e.printStackTrace();
         }
         try {
-            Assert.assertTrue(BQScrollableResultSetFunctionTest.con.isClosed());
+            Assert.assertTrue(BQResultSetFunctionTest.con.isClosed());
         } catch (SQLException e1) {
             e1.printStackTrace();
         }
 
         try {
-            BQScrollableResultSetFunctionTest.con.isValid(0);
+            BQResultSetFunctionTest.con.isValid(0);
         } catch (SQLException e) {
             Assert.assertTrue(true);
             e.printStackTrace();
@@ -259,8 +255,8 @@ public class BQScrollableResultSetFunctionTest {
     public void NewConnection() {
 
         try {
-            if (BQScrollableResultSetFunctionTest.con == null
-                    || !BQScrollableResultSetFunctionTest.con.isValid(0)) {
+            if (BQResultSetFunctionTest.con == null
+                    || !BQResultSetFunctionTest.con.isValid(0)) {
                 this.logger.info("Testing the JDBC driver");
                 try {
                     Class.forName("net.starschema.clouddb.jdbc.BQDriver");
@@ -268,7 +264,7 @@ public class BQScrollableResultSetFunctionTest {
                             .constructUrlFromPropertiesFile(BQSupportFuncts
                                     .readFromPropFile(getClass().getResource("/installedaccount1.properties").getFile()));
                     jdbcUrl += "&useLegacySql=true";
-                    BQScrollableResultSetFunctionTest.con = DriverManager
+                    BQResultSetFunctionTest.con = DriverManager
                             .getConnection(jdbcUrl,
                                     BQSupportFuncts
                                             .readFromPropFile(getClass().getResource("/installedaccount1.properties").getFile()));
@@ -277,7 +273,7 @@ public class BQScrollableResultSetFunctionTest {
                     this.logger.error("Error in connection" + e.toString());
                     Assert.fail("General Exception:" + e.toString());
                 }
-                this.logger.info(((BQConnection) BQScrollableResultSetFunctionTest.con)
+                this.logger.info(((BQConnection) BQResultSetFunctionTest.con)
                         .getURLPART());
             }
         } catch (SQLException e) {
@@ -288,34 +284,55 @@ public class BQScrollableResultSetFunctionTest {
 
     // Comprehensive Tests:
 
+    /**
+     * Prints a String[][] QueryResult to Log
+     *
+     * @param input
+     */
+    private void printer(String[][] input) {
+        for (int s = 0; s < input[0].length; s++) {
+            String Output = "";
+            for (int i = 0; i < input.length; i++) {
+                if (i == input.length - 1) {
+                    Output += input[i][s];
+                } else {
+                    Output += input[i][s] + "\t";
+                }
+            }
+            this.logger.debug(Output);
+        }
+    }
+
     public void QueryLoad() {
         final String sql = "SELECT TOP(word,10) AS word, COUNT(*) as count FROM publicdata:samples.shakespeare";
         final String description = "The top 10 word from shakespeare #TOP #COUNT";
         String[][] expectation = new String[][]{
-                { "you", "yet", "would", "world", "without", "with", "will", "why", "whose", "whom" },
-                { "42", "42", "42", "42", "42", "42", "42", "42", "42", "42" } };
+         { "you", "yet", "would", "world", "without", "with", "will",
+         "why", "whose", "whom" },
+         { "42", "42", "42", "42", "42", "42", "42", "42", "42", "42" } };
+
         this.logger.info("Test number: 01");
         this.logger.info("Running query:" + sql);
 
         try {
-            Statement stmt = BQScrollableResultSetFunctionTest.con
-                    .createStatement(ResultSet.TYPE_SCROLL_INSENSITIVE, ResultSet.CONCUR_READ_ONLY);
+            Statement stmt = BQResultSetFunctionTest.con.createStatement(
+                    ResultSet.TYPE_SCROLL_INSENSITIVE, ResultSet.CONCUR_READ_ONLY);
             stmt.setQueryTimeout(500);
-            BQScrollableResultSetFunctionTest.Result = stmt.executeQuery(sql);
+            BQResultSetFunctionTest.Result = stmt.executeQuery(sql);
         } catch (SQLException e) {
             this.logger.error("SQLexception" + e.toString());
             Assert.fail("SQLException" + e.toString());
         }
-        Assert.assertNotNull(BQScrollableResultSetFunctionTest.Result);
+        Assert.assertNotNull(BQResultSetFunctionTest.Result);
 
         this.logger.debug(description);
-        HelperFunctions.printer(expectation);
+        this.printer(expectation);
 
 
         try {
             Assert.assertTrue("Comparing failed in the String[][] array", this
                     .comparer(expectation, BQSupportMethods
-                            .GetQueryResult(BQScrollableResultSetFunctionTest.Result)));
+                            .GetQueryResult(BQResultSetFunctionTest.Result)));
         } catch (SQLException e) {
             this.logger.error("SQLexception" + e.toString());
             Assert.fail(e.toString());
@@ -325,9 +342,9 @@ public class BQScrollableResultSetFunctionTest {
     @Test
     public void ResultSetMetadata() {
         try {
-            this.logger.debug(BQScrollableResultSetFunctionTest.Result.getMetaData()
+            this.logger.debug(BQResultSetFunctionTest.Result.getMetaData()
                     .getSchemaName(1));
-            this.logger.debug("{}", BQScrollableResultSetFunctionTest.Result.getMetaData()
+            this.logger.debug("{}", BQResultSetFunctionTest.Result.getMetaData()
                     .getScale(1));
         } catch (SQLException e) {
             this.logger.error("SQLexception" + e.toString());
@@ -338,7 +355,7 @@ public class BQScrollableResultSetFunctionTest {
     @Test
     public void TestResultIndexOutofBound() {
         try {
-            this.logger.debug("{}", BQScrollableResultSetFunctionTest.Result.getBoolean(99));
+            this.logger.debug("{}", BQResultSetFunctionTest.Result.getBoolean(99));
         } catch (SQLException e) {
             Assert.assertTrue(true);
             this.logger.error("SQLexception" + e.toString());
@@ -348,44 +365,44 @@ public class BQScrollableResultSetFunctionTest {
     @Test
     public void TestResultSetAbsolute() {
         try {
-            Assert.assertTrue(BQScrollableResultSetFunctionTest.Result.absolute(1));
+            Assert.assertTrue(BQResultSetFunctionTest.Result.absolute(1));
             Assert.assertEquals("you",
-                    BQScrollableResultSetFunctionTest.Result.getString(1));
-            Assert.assertTrue(BQScrollableResultSetFunctionTest.Result.absolute(2));
+                    BQResultSetFunctionTest.Result.getString(1));
+            Assert.assertTrue(BQResultSetFunctionTest.Result.absolute(2));
             Assert.assertEquals("yet",
-                    BQScrollableResultSetFunctionTest.Result.getString(1));
-            Assert.assertTrue(BQScrollableResultSetFunctionTest.Result.absolute(3));
+                    BQResultSetFunctionTest.Result.getString(1));
+            Assert.assertTrue(BQResultSetFunctionTest.Result.absolute(3));
             Assert.assertEquals("would",
-                    BQScrollableResultSetFunctionTest.Result.getString(1));
-            Assert.assertTrue(BQScrollableResultSetFunctionTest.Result.absolute(4));
+                    BQResultSetFunctionTest.Result.getString(1));
+            Assert.assertTrue(BQResultSetFunctionTest.Result.absolute(4));
             Assert.assertEquals("world",
-                    BQScrollableResultSetFunctionTest.Result.getString(1));
-            Assert.assertTrue(BQScrollableResultSetFunctionTest.Result.absolute(5));
+                    BQResultSetFunctionTest.Result.getString(1));
+            Assert.assertTrue(BQResultSetFunctionTest.Result.absolute(5));
             Assert.assertEquals("without",
-                    BQScrollableResultSetFunctionTest.Result.getString(1));
-            Assert.assertTrue(BQScrollableResultSetFunctionTest.Result.absolute(6));
+                    BQResultSetFunctionTest.Result.getString(1));
+            Assert.assertTrue(BQResultSetFunctionTest.Result.absolute(6));
             Assert.assertEquals("with",
-                    BQScrollableResultSetFunctionTest.Result.getString(1));
-            Assert.assertTrue(BQScrollableResultSetFunctionTest.Result.absolute(7));
+                    BQResultSetFunctionTest.Result.getString(1));
+            Assert.assertTrue(BQResultSetFunctionTest.Result.absolute(7));
             Assert.assertEquals("will",
-                    BQScrollableResultSetFunctionTest.Result.getString(1));
-            Assert.assertTrue(BQScrollableResultSetFunctionTest.Result.absolute(8));
+                    BQResultSetFunctionTest.Result.getString(1));
+            Assert.assertTrue(BQResultSetFunctionTest.Result.absolute(8));
             Assert.assertEquals("why",
-                    BQScrollableResultSetFunctionTest.Result.getString(1));
-            Assert.assertTrue(BQScrollableResultSetFunctionTest.Result.absolute(9));
+                    BQResultSetFunctionTest.Result.getString(1));
+            Assert.assertTrue(BQResultSetFunctionTest.Result.absolute(9));
             Assert.assertEquals("whose",
-                    BQScrollableResultSetFunctionTest.Result.getString(1));
-            Assert.assertTrue(BQScrollableResultSetFunctionTest.Result.absolute(10));
+                    BQResultSetFunctionTest.Result.getString(1));
+            Assert.assertTrue(BQResultSetFunctionTest.Result.absolute(10));
             Assert.assertEquals("whom",
-                    BQScrollableResultSetFunctionTest.Result.getString(1));
+                    BQResultSetFunctionTest.Result.getString(1));
         } catch (SQLException e) {
             this.logger.error("SQLexception" + e.toString());
             Assert.fail("SQLException" + e.toString());
         }
 
         try {
-            Assert.assertFalse(BQScrollableResultSetFunctionTest.Result.absolute(0));
-            Assert.assertEquals("", BQScrollableResultSetFunctionTest.Result.getString(1));
+            Assert.assertFalse(BQResultSetFunctionTest.Result.absolute(0));
+            Assert.assertEquals("", BQResultSetFunctionTest.Result.getString(1));
         } catch (SQLException e) {
             boolean ct = e.toString().contains(
                     "Cursor is not in a valid Position");
@@ -398,8 +415,8 @@ public class BQScrollableResultSetFunctionTest {
         }
 
         try {
-            Assert.assertFalse(BQScrollableResultSetFunctionTest.Result.absolute(11));
-            Assert.assertEquals("", BQScrollableResultSetFunctionTest.Result.getString(1));
+            Assert.assertFalse(BQResultSetFunctionTest.Result.absolute(11));
+            Assert.assertEquals("", BQResultSetFunctionTest.Result.getString(1));
         } catch (SQLException e) {
             boolean ct = e.toString().contains(
                     "Cursor is not in a valid Position");
@@ -415,18 +432,18 @@ public class BQScrollableResultSetFunctionTest {
     @Test
     public void TestResultSetAfterlast() {
         try {
-            BQScrollableResultSetFunctionTest.Result.afterLast();
-            Assert.assertTrue(BQScrollableResultSetFunctionTest.Result.previous());
+            BQResultSetFunctionTest.Result.afterLast();
+            Assert.assertTrue(BQResultSetFunctionTest.Result.previous());
             Assert.assertEquals("whom",
-                    BQScrollableResultSetFunctionTest.Result.getString(1));
+                    BQResultSetFunctionTest.Result.getString(1));
         } catch (SQLException e) {
             this.logger.error("SQLexception" + e.toString());
             Assert.fail("SQLException" + e.toString());
         }
 
         try {
-            BQScrollableResultSetFunctionTest.Result.afterLast();
-            Assert.assertEquals("", BQScrollableResultSetFunctionTest.Result.getString(1));
+            BQResultSetFunctionTest.Result.afterLast();
+            Assert.assertEquals("", BQResultSetFunctionTest.Result.getString(1));
         } catch (SQLException e) {
             boolean ct = e.toString().contains(
                     "Cursor is not in a valid Position");
@@ -442,18 +459,18 @@ public class BQScrollableResultSetFunctionTest {
     @Test
     public void TestResultSetBeforeFirst() {
         try {
-            BQScrollableResultSetFunctionTest.Result.beforeFirst();
-            Assert.assertTrue(BQScrollableResultSetFunctionTest.Result.next());
+            BQResultSetFunctionTest.Result.beforeFirst();
+            Assert.assertTrue(BQResultSetFunctionTest.Result.next());
             Assert.assertEquals("you",
-                    BQScrollableResultSetFunctionTest.Result.getString(1));
+                    BQResultSetFunctionTest.Result.getString(1));
         } catch (SQLException e) {
             this.logger.error("SQLexception" + e.toString());
             Assert.fail("SQLException" + e.toString());
         }
 
         try {
-            BQScrollableResultSetFunctionTest.Result.beforeFirst();
-            Assert.assertEquals("", BQScrollableResultSetFunctionTest.Result.getString(1));
+            BQResultSetFunctionTest.Result.beforeFirst();
+            Assert.assertEquals("", BQResultSetFunctionTest.Result.getString(1));
         } catch (SQLException e) {
             boolean ct = e.toString().contains(
                     "Cursor is not in a valid Position");
@@ -469,8 +486,8 @@ public class BQScrollableResultSetFunctionTest {
     @Test
     public void TestResultSetFirst() {
         try {
-            Assert.assertTrue(BQScrollableResultSetFunctionTest.Result.first());
-            Assert.assertTrue(BQScrollableResultSetFunctionTest.Result.isFirst());
+            Assert.assertTrue(BQResultSetFunctionTest.Result.first());
+            Assert.assertTrue(BQResultSetFunctionTest.Result.isFirst());
         } catch (SQLException e) {
             this.logger.error("SQLexception" + e.toString());
             Assert.fail("SQLException" + e.toString());
@@ -480,9 +497,9 @@ public class BQScrollableResultSetFunctionTest {
     @Test
     public void TestResultSetgetBoolean() {
         try {
-            Assert.assertTrue(BQScrollableResultSetFunctionTest.Result.absolute(1));
+            Assert.assertTrue(BQResultSetFunctionTest.Result.absolute(1));
             Assert.assertEquals(Boolean.parseBoolean("42"),
-                    BQScrollableResultSetFunctionTest.Result.getBoolean(2));
+                    BQResultSetFunctionTest.Result.getBoolean(2));
         } catch (SQLException e) {
             this.logger.error("SQLexception" + e.toString());
             Assert.fail("SQLException" + e.toString());
@@ -492,9 +509,9 @@ public class BQScrollableResultSetFunctionTest {
     @Test
     public void TestResultSetgetFloat() {
         try {
-            Assert.assertTrue(BQScrollableResultSetFunctionTest.Result.absolute(1));
+            Assert.assertTrue(BQResultSetFunctionTest.Result.absolute(1));
             Assert.assertEquals(new Float(42),
-                    BQScrollableResultSetFunctionTest.Result.getFloat(2));
+                    BQResultSetFunctionTest.Result.getFloat(2));
         } catch (SQLException e) {
             this.logger.error("SQLexception" + e.toString());
             Assert.fail("SQLException" + e.toString());
@@ -504,8 +521,8 @@ public class BQScrollableResultSetFunctionTest {
     @Test
     public void TestResultSetgetInteger() {
         try {
-            Assert.assertTrue(BQScrollableResultSetFunctionTest.Result.absolute(1));
-            Assert.assertEquals(42, BQScrollableResultSetFunctionTest.Result.getInt(2));
+            Assert.assertTrue(BQResultSetFunctionTest.Result.absolute(1));
+            Assert.assertEquals(42, BQResultSetFunctionTest.Result.getInt(2));
         } catch (SQLException e) {
             this.logger.error("SQLexception" + e.toString());
             Assert.fail("SQLException" + e.toString());
@@ -516,19 +533,19 @@ public class BQScrollableResultSetFunctionTest {
     public void TestResultSetgetRow() {
 
         try {
-            Assert.assertTrue(BQScrollableResultSetFunctionTest.Result.absolute(1));
-            Assert.assertEquals(1, BQScrollableResultSetFunctionTest.Result.getRow());
-            Assert.assertTrue(BQScrollableResultSetFunctionTest.Result.absolute(10));
-            Assert.assertEquals(10, BQScrollableResultSetFunctionTest.Result.getRow());
+            Assert.assertTrue(BQResultSetFunctionTest.Result.absolute(1));
+            Assert.assertEquals(1, BQResultSetFunctionTest.Result.getRow());
+            Assert.assertTrue(BQResultSetFunctionTest.Result.absolute(10));
+            Assert.assertEquals(10, BQResultSetFunctionTest.Result.getRow());
         } catch (SQLException e) {
             this.logger.error("SQLexception" + e.toString());
             Assert.fail("SQLException" + e.toString());
         }
         try {
-            BQScrollableResultSetFunctionTest.Result.beforeFirst();
-            Assert.assertEquals(0, BQScrollableResultSetFunctionTest.Result.getRow());
-            BQScrollableResultSetFunctionTest.Result.afterLast();
-            Assert.assertEquals(0, BQScrollableResultSetFunctionTest.Result.getRow());
+            BQResultSetFunctionTest.Result.beforeFirst();
+            Assert.assertEquals(0, BQResultSetFunctionTest.Result.getRow());
+            BQResultSetFunctionTest.Result.afterLast();
+            Assert.assertEquals(0, BQResultSetFunctionTest.Result.getRow());
         } catch (SQLException e) {
             this.logger.error("SQLexception" + e.toString());
             Assert.fail("SQLException" + e.toString());
@@ -538,12 +555,12 @@ public class BQScrollableResultSetFunctionTest {
     @Test
     public void TestResultSetgetString() {
         try {
-            Assert.assertTrue(BQScrollableResultSetFunctionTest.Result.first());
+            Assert.assertTrue(BQResultSetFunctionTest.Result.first());
             Assert.assertEquals("you",
-                    BQScrollableResultSetFunctionTest.Result.getString(1));
-            Assert.assertTrue(BQScrollableResultSetFunctionTest.Result.last());
+                    BQResultSetFunctionTest.Result.getString(1));
+            Assert.assertTrue(BQResultSetFunctionTest.Result.last());
             Assert.assertEquals("whom",
-                    BQScrollableResultSetFunctionTest.Result.getString(1));
+                    BQResultSetFunctionTest.Result.getString(1));
         } catch (SQLException e) {
             this.logger.error("SQLexception" + e.toString());
             Assert.fail("SQLException" + e.toString());
@@ -553,8 +570,8 @@ public class BQScrollableResultSetFunctionTest {
     @Test
     public void TestResultSetLast() {
         try {
-            Assert.assertTrue(BQScrollableResultSetFunctionTest.Result.last());
-            Assert.assertTrue(BQScrollableResultSetFunctionTest.Result.isLast());
+            Assert.assertTrue(BQResultSetFunctionTest.Result.last());
+            Assert.assertTrue(BQResultSetFunctionTest.Result.isLast());
         } catch (SQLException e) {
             this.logger.error("SQLexception" + e.toString());
             Assert.fail("SQLException" + e.toString());
@@ -564,42 +581,42 @@ public class BQScrollableResultSetFunctionTest {
     @Test
     public void TestResultSetNext() {
         try {
-            Assert.assertTrue(BQScrollableResultSetFunctionTest.Result.first());
-            Assert.assertTrue(BQScrollableResultSetFunctionTest.Result.next());
+            Assert.assertTrue(BQResultSetFunctionTest.Result.first());
+            Assert.assertTrue(BQResultSetFunctionTest.Result.next());
             Assert.assertEquals("yet",
-                    BQScrollableResultSetFunctionTest.Result.getString(1));
-            Assert.assertTrue(BQScrollableResultSetFunctionTest.Result.next());
+                    BQResultSetFunctionTest.Result.getString(1));
+            Assert.assertTrue(BQResultSetFunctionTest.Result.next());
             Assert.assertEquals("would",
-                    BQScrollableResultSetFunctionTest.Result.getString(1));
-            Assert.assertTrue(BQScrollableResultSetFunctionTest.Result.next());
+                    BQResultSetFunctionTest.Result.getString(1));
+            Assert.assertTrue(BQResultSetFunctionTest.Result.next());
             Assert.assertEquals("world",
-                    BQScrollableResultSetFunctionTest.Result.getString(1));
-            Assert.assertTrue(BQScrollableResultSetFunctionTest.Result.next());
+                    BQResultSetFunctionTest.Result.getString(1));
+            Assert.assertTrue(BQResultSetFunctionTest.Result.next());
             Assert.assertEquals("without",
-                    BQScrollableResultSetFunctionTest.Result.getString(1));
-            Assert.assertTrue(BQScrollableResultSetFunctionTest.Result.next());
+                    BQResultSetFunctionTest.Result.getString(1));
+            Assert.assertTrue(BQResultSetFunctionTest.Result.next());
             Assert.assertEquals("with",
-                    BQScrollableResultSetFunctionTest.Result.getString(1));
-            Assert.assertTrue(BQScrollableResultSetFunctionTest.Result.next());
+                    BQResultSetFunctionTest.Result.getString(1));
+            Assert.assertTrue(BQResultSetFunctionTest.Result.next());
             Assert.assertEquals("will",
-                    BQScrollableResultSetFunctionTest.Result.getString(1));
-            Assert.assertTrue(BQScrollableResultSetFunctionTest.Result.next());
+                    BQResultSetFunctionTest.Result.getString(1));
+            Assert.assertTrue(BQResultSetFunctionTest.Result.next());
             Assert.assertEquals("why",
-                    BQScrollableResultSetFunctionTest.Result.getString(1));
-            Assert.assertTrue(BQScrollableResultSetFunctionTest.Result.next());
+                    BQResultSetFunctionTest.Result.getString(1));
+            Assert.assertTrue(BQResultSetFunctionTest.Result.next());
             Assert.assertEquals("whose",
-                    BQScrollableResultSetFunctionTest.Result.getString(1));
-            Assert.assertTrue(BQScrollableResultSetFunctionTest.Result.next());
+                    BQResultSetFunctionTest.Result.getString(1));
+            Assert.assertTrue(BQResultSetFunctionTest.Result.next());
             Assert.assertEquals("whom",
-                    BQScrollableResultSetFunctionTest.Result.getString(1));
-            Assert.assertFalse(BQScrollableResultSetFunctionTest.Result.next());
+                    BQResultSetFunctionTest.Result.getString(1));
+            Assert.assertFalse(BQResultSetFunctionTest.Result.next());
         } catch (SQLException e) {
             this.logger.error("SQLexception" + e.toString());
             Assert.fail("SQLException" + e.toString());
         }
 
         try {
-            Assert.assertEquals("", BQScrollableResultSetFunctionTest.Result.getString(1));
+            Assert.assertEquals("", BQResultSetFunctionTest.Result.getString(1));
         } catch (SQLException e) {
             boolean ct = e.toString().contains(
                     "Cursor is not in a valid Position");
@@ -615,41 +632,41 @@ public class BQScrollableResultSetFunctionTest {
     @Test
     public void TestResultSetPrevious() {
         try {
-            Assert.assertTrue(BQScrollableResultSetFunctionTest.Result.last());
-            Assert.assertTrue(BQScrollableResultSetFunctionTest.Result.previous());
+            Assert.assertTrue(BQResultSetFunctionTest.Result.last());
+            Assert.assertTrue(BQResultSetFunctionTest.Result.previous());
             Assert.assertEquals("whose",
-                    BQScrollableResultSetFunctionTest.Result.getString(1));
-            Assert.assertTrue(BQScrollableResultSetFunctionTest.Result.previous());
+                    BQResultSetFunctionTest.Result.getString(1));
+            Assert.assertTrue(BQResultSetFunctionTest.Result.previous());
             Assert.assertEquals("why",
-                    BQScrollableResultSetFunctionTest.Result.getString(1));
-            Assert.assertTrue(BQScrollableResultSetFunctionTest.Result.previous());
+                    BQResultSetFunctionTest.Result.getString(1));
+            Assert.assertTrue(BQResultSetFunctionTest.Result.previous());
             Assert.assertEquals("will",
-                    BQScrollableResultSetFunctionTest.Result.getString(1));
-            Assert.assertTrue(BQScrollableResultSetFunctionTest.Result.previous());
+                    BQResultSetFunctionTest.Result.getString(1));
+            Assert.assertTrue(BQResultSetFunctionTest.Result.previous());
             Assert.assertEquals("with",
-                    BQScrollableResultSetFunctionTest.Result.getString(1));
-            Assert.assertTrue(BQScrollableResultSetFunctionTest.Result.previous());
+                    BQResultSetFunctionTest.Result.getString(1));
+            Assert.assertTrue(BQResultSetFunctionTest.Result.previous());
             Assert.assertEquals("without",
-                    BQScrollableResultSetFunctionTest.Result.getString(1));
-            Assert.assertTrue(BQScrollableResultSetFunctionTest.Result.previous());
+                    BQResultSetFunctionTest.Result.getString(1));
+            Assert.assertTrue(BQResultSetFunctionTest.Result.previous());
             Assert.assertEquals("world",
-                    BQScrollableResultSetFunctionTest.Result.getString(1));
-            Assert.assertTrue(BQScrollableResultSetFunctionTest.Result.previous());
+                    BQResultSetFunctionTest.Result.getString(1));
+            Assert.assertTrue(BQResultSetFunctionTest.Result.previous());
             Assert.assertEquals("would",
-                    BQScrollableResultSetFunctionTest.Result.getString(1));
-            Assert.assertTrue(BQScrollableResultSetFunctionTest.Result.previous());
+                    BQResultSetFunctionTest.Result.getString(1));
+            Assert.assertTrue(BQResultSetFunctionTest.Result.previous());
             Assert.assertEquals("yet",
-                    BQScrollableResultSetFunctionTest.Result.getString(1));
-            Assert.assertTrue(BQScrollableResultSetFunctionTest.Result.previous());
+                    BQResultSetFunctionTest.Result.getString(1));
+            Assert.assertTrue(BQResultSetFunctionTest.Result.previous());
             Assert.assertEquals("you",
-                    BQScrollableResultSetFunctionTest.Result.getString(1));
-            Assert.assertFalse(BQScrollableResultSetFunctionTest.Result.previous());
+                    BQResultSetFunctionTest.Result.getString(1));
+            Assert.assertFalse(BQResultSetFunctionTest.Result.previous());
         } catch (SQLException e) {
             this.logger.error("SQLexception" + e.toString());
             Assert.fail("SQLException" + e.toString());
         }
         try {
-            Assert.assertEquals("", BQScrollableResultSetFunctionTest.Result.getString(1));
+            Assert.assertEquals("", BQResultSetFunctionTest.Result.getString(1));
         } catch (SQLException e) {
             boolean ct = e.toString().contains(
                     "Cursor is not in a valid Position");
@@ -665,35 +682,35 @@ public class BQScrollableResultSetFunctionTest {
     @Test
     public void TestResultSetRelative() {
         try {
-            Assert.assertTrue(BQScrollableResultSetFunctionTest.Result.absolute(1));
+            Assert.assertTrue(BQResultSetFunctionTest.Result.absolute(1));
             Assert.assertEquals("you",
-                    BQScrollableResultSetFunctionTest.Result.getString(1));
-            Assert.assertTrue(BQScrollableResultSetFunctionTest.Result.relative(1));
+                    BQResultSetFunctionTest.Result.getString(1));
+            Assert.assertTrue(BQResultSetFunctionTest.Result.relative(1));
             Assert.assertEquals("yet",
-                    BQScrollableResultSetFunctionTest.Result.getString(1));
-            Assert.assertTrue(BQScrollableResultSetFunctionTest.Result.relative(2));
+                    BQResultSetFunctionTest.Result.getString(1));
+            Assert.assertTrue(BQResultSetFunctionTest.Result.relative(2));
             Assert.assertEquals("world",
-                    BQScrollableResultSetFunctionTest.Result.getString(1));
-            Assert.assertTrue(BQScrollableResultSetFunctionTest.Result.relative(5));
+                    BQResultSetFunctionTest.Result.getString(1));
+            Assert.assertTrue(BQResultSetFunctionTest.Result.relative(5));
             Assert.assertEquals("whose",
-                    BQScrollableResultSetFunctionTest.Result.getString(1));
-            Assert.assertTrue(BQScrollableResultSetFunctionTest.Result.relative(-5));
+                    BQResultSetFunctionTest.Result.getString(1));
+            Assert.assertTrue(BQResultSetFunctionTest.Result.relative(-5));
             Assert.assertEquals("world",
-                    BQScrollableResultSetFunctionTest.Result.getString(1));
-            Assert.assertTrue(BQScrollableResultSetFunctionTest.Result.relative(-2));
+                    BQResultSetFunctionTest.Result.getString(1));
+            Assert.assertTrue(BQResultSetFunctionTest.Result.relative(-2));
             Assert.assertEquals("yet",
-                    BQScrollableResultSetFunctionTest.Result.getString(1));
-            Assert.assertTrue(BQScrollableResultSetFunctionTest.Result.relative(-1));
+                    BQResultSetFunctionTest.Result.getString(1));
+            Assert.assertTrue(BQResultSetFunctionTest.Result.relative(-1));
             Assert.assertEquals("you",
-                    BQScrollableResultSetFunctionTest.Result.getString(1));
+                    BQResultSetFunctionTest.Result.getString(1));
         } catch (SQLException e) {
             this.logger.error("SQLexception" + e.toString());
             Assert.fail("SQLException" + e.toString());
         }
         try {
-            Assert.assertTrue(BQScrollableResultSetFunctionTest.Result.first());
-            Assert.assertFalse(BQScrollableResultSetFunctionTest.Result.relative(-1));
-            Assert.assertEquals("", BQScrollableResultSetFunctionTest.Result.getString(1));
+            Assert.assertTrue(BQResultSetFunctionTest.Result.first());
+            Assert.assertFalse(BQResultSetFunctionTest.Result.relative(-1));
+            Assert.assertEquals("", BQResultSetFunctionTest.Result.getString(1));
         } catch (SQLException e) {
             boolean ct = e.toString().contains(
                     "Cursor is not in a valid Position");
@@ -706,9 +723,9 @@ public class BQScrollableResultSetFunctionTest {
         }
 
         try {
-            Assert.assertTrue(BQScrollableResultSetFunctionTest.Result.last());
-            Assert.assertFalse(BQScrollableResultSetFunctionTest.Result.relative(1));
-            Assert.assertEquals("", BQScrollableResultSetFunctionTest.Result.getString(1));
+            Assert.assertTrue(BQResultSetFunctionTest.Result.last());
+            Assert.assertFalse(BQResultSetFunctionTest.Result.relative(1));
+            Assert.assertEquals("", BQResultSetFunctionTest.Result.getString(1));
         } catch (SQLException e) {
             boolean ct = e.toString().contains(
                     "Cursor is not in a valid Position");
