+++ src/test/java/net/starschema/clouddb/jdbc/PreparedStatementTests.java
@@ -22,7 +22,7 @@
  * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
  * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
-package BQJDBC.QueryResultTest;
+package net.starschema.clouddb.jdbc;
 
 import java.io.StringReader;
 import java.math.BigDecimal;
@@ -36,8 +36,6 @@ import java.text.ParseException;
 import java.text.SimpleDateFormat;
 
 import junit.framework.Assert;
-import net.starschema.clouddb.jdbc.BQSupportFuncts;
-import net.starschema.clouddb.jdbc.BQSupportMethods;
 
 import org.junit.Before;
 import org.junit.Test;
