+++ src/test/java/net/starschema/clouddb/jdbc/QueryResultTest.java
