+++ src/test/java/net/starschema/clouddb/jdbc/Tokens.java
@@ -22,7 +22,7 @@
  * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
  * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
-package BQJDBC.QueryResultTest;
+package net.starschema.clouddb.jdbc;
 
 public enum Tokens {
     EXPRESSION("EXPRESSION"),
